+++ lofty/tests/files/aac.rs
@@ -44,7 +44,7 @@ fn read_with_junk_bytes_between_frames() {
 	assert_eq!(id3v2_tag.album().as_deref(), Some("album test"));
 	assert_eq!(id3v2_tag.title().as_deref(), Some("title test"));
 	assert_eq!(
-		id3v2_tag.get_string(&ItemKey::EncoderSettings),
+		id3v2_tag.get_string(ItemKey::EncoderSettings),
 		Some("Lavf58.62.100")
 	);